move SyncType stuff to z_coin_activation_prefs
